requirements.txt
streamlit>=1.32.0
openai>=1.14.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.20.0
python-dotenv>=1.0.0
flask>=3.0.0
feedparser>=6.0.0
wordcloud>=1.9.3
matplotlib>=3.8.0
pillow>=10.2.0
